First playable version - no bugs found
Changes to how the computer randomly selects and how the
playRound() function is called via button-clicks. Button-click function
calls was remade as callback for them to function.
The newGame() function now resets the UI correct without saving the
scores internally.
The game declares a winner when one player reaches the value of
the pointsToWin variable.